fix webpack config process optimize
