+++ example/tasks/config/client/webpack.config.ts
@@ -1,10 +1,10 @@
-import { generateWebpackConfig } from 'bistrio'
+import { generateClientWebpackConfig } from 'bistrio'
 import { config } from '../../config/index.ts'
 import { fileURLToPath } from 'node:url'
 
 const __filename = fileURLToPath(import.meta.url)
 
-const webpackConfig = generateWebpackConfig({
+const webpackConfig = generateClientWebpackConfig({
   config,
   bundlerConfigPath: __filename,
 })
