+++ example/tasks/config/server/tsconfig.server.production.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../tsconfig.json"
+}
