+++ package.json
@@ -44,6 +44,7 @@
     "express": "~4.18.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "webpack-node-externals": "^3.0.0",
     "zod": "^3.22.4"
   },
   "devDependencies": {
@@ -53,6 +54,7 @@
     "@types/jest": "^29.5.1",
     "@types/node": "^18.11.18",
     "@types/react-dom": "^18.0.10",
+    "@types/webpack-node-externals": "^3.0.4",
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
     "eslint": "^8.32.0",
