+++ src/lib/filemap.ts
@@ -1,0 +1,3 @@
+export type Filemap = {
+  js: { [key: string]: string }
+}
